Add basic (and inconsistent) knock-back to player on damage.
There seems to be some inconsistency in the magnitude of
knock-back each time the player collides with an enemy.
For example, there seems to be more force applied on the X
axis while the player is grounded.  Another example is when
the player sits on top of an enemy; the first collision
yields a smaller amount of upward force than subsequent
damage knock-back.

Even though this isn't perfect, I'm going to commit it
because 1) it's good enough for now and 2) it will allow
more exposure to developers and have a greater chance of
being fixed.

Signed-off-by: suicidepills <atom.mac@gmail.com>